re-arrange tests and permissions so reports app is completely removable
