+++ dashboard/apps/dashboard/fixtures/auth_group_permissions.yaml
@@ -26,10 +26,6 @@
       - ['change_skill', 'dashboard', 'skill']
       - ['change_product', 'dashboard', 'product']
       - ['export_person_rates', 'dashboard', 'person']
-      - ['upload_payroll', 'reports', 'productproxy']
-      - ['adjustment_export', 'reports', 'productproxy']
-      - ['intercompany_export', 'reports', 'productproxy']
-      - ['productdetail_export', 'reports', 'productproxy']
       - ['change_task', 'dashboard', 'task']
       - ['change_cost', 'dashboard', 'cost']
       - ['add_cost', 'dashboard', 'cost']
@@ -52,6 +48,7 @@
       - ['change_link', 'dashboard', 'link']
       - ['add_link', 'dashboard', 'link']
       - ['delete_link', 'dashboard', 'link']
+      - ['financial_access', 'dashboard', 'product']
 - model: auth.group
   pk: 3
   fields:
